Fix an out-of-bounds error on sockaddrs.
The `invalidAddresses` vector in `client.c` is used
to hold invalid addresses `binkd` should not use.
However, the array was of type `struct sockaddr`,
which is not large enough to hold all of the
protocol-specific data of the `struct sockaddr_*`
structures.  As a result, `binkd` would access
out-of-bounds memory when examining elements of
the array.

Fixed by redefining `invalidAddresses` to be an
array of type `struct sockaddr_storage`, which is
guaranteed by POSIX to be large enough to hold all
data associated with a socket address for any
protocol family.

Signed-off-by: Dan Cross <cross@fat-dragon.org>